requirements.txt
# Data acquisition & scripting
requests>=2.31.0
python-dotenv>=1.0.0

# Data Analysis
pandas>=2.1.0
numpy>=1.26.0

# AI/ML Pipeline
scikit-learn>=1.3.0
matplotlib>=3.8.0

# Database
pymongo[srv]>=4.5.0
dnspython>=2.4.0
